refactor(licensing): split claim-aware resolution out of apps.core
apps.core.licensing previously mixed pure license policy with claim/source
license resolution, requiring a TYPE_CHECKING import of Claim and a runtime
lazy import of SourceFieldLicense to break a core→provenance cycle.

- Move SourceFieldLicenseMap, build_source_field_license_map, and
  resolve_effective_license to apps.provenance.licensing.
- Move IMAGE_FIELDS (catalog/ingest-specific) into apps.catalog.resolve.
- Move the CONTENT_DISPLAY_POLICY cache-invalidation listener from
  CoreConfig.ready() to apps.catalog.signals; core no longer imports
  apps.catalog.cache. Adds tests for the moved listener.
- Drop now-unnecessary TYPE_CHECKING blocks and function-body lazy
  imports across apps.core (licensing, markdown, markdown_links, apps).

After: apps.core has zero outbound apps.* imports outside tests/migrations.
